jestem gloda, ide jesc wiec daje nazwe na szybko commita i robie to byle jak
