--- src/main/java/common/commonbackend/tasks/TaskPriceUpdateAlgorithm.java
@@ -1,8 +1,0 @@
-package common.commonbackend.tasks;
-
-import java.time.LocalDate;
-import java.time.Period;
-
-public interface TaskPriceUpdateAlgorithm {
-    long getNewPrice(long price, LocalDate lastDoneDate, Period period);
-}
