+++ src/main/java/common/commonbackend/tasks/updatealgorithms/TaskPriceUpdateAlgorithm.java
@@ -1,0 +1,8 @@
+package common.commonbackend.tasks.updatealgorithms;
+
+import java.time.LocalDate;
+import java.time.Period;
+
+interface TaskPriceUpdateAlgorithm {
+    long getNewPrice(long oldPrice, LocalDate lastDoneDate, Period repetitionRate);
+}
