+++ src/test/java/common/commonbackend/tasks/updatealgorithms/TaskPriceUpdaterServiceTest.java
@@ -1,6 +1,7 @@
-package common.commonbackend.tasks;
+package common.commonbackend.tasks.updatealgorithms;
 
 import common.commonbackend.rooms.Room;
+import common.commonbackend.tasks.Task;
 import org.junit.jupiter.api.Test;
 import org.junit.runner.RunWith;
 import org.mockito.Mock;
